课程模块update

+++ attendance-dao/src/main/java/com/jj/stu/attendance/dao/response/course/PageCourseListResponse.java
@@ -8,5 +8,5 @@ import java.util.List;
 @Data
 public class PageCourseListResponse {
     List<PageCourseDto> pageCourseDtoList;
-    Integer totalSize;
+    Long totalSize;
 }
