+++ attendance-dao/src/main/java/com/jj/stu/attendance/dao/response/selectedCourse/PageSelectedCourseResponse.java
@@ -12,5 +12,5 @@ import java.util.List;
 public class PageSelectedCourseResponse {
     List<SelectedCourse> selectedCourseList;
 
-    Integer totalSize;
+    Long totalSize;
 }
